Refacto generate_grid_sdf into its own module
