+++ mesh_to_sdf/src/generate/mod.rs
@@ -1,0 +1,3 @@
+//! Module for generating SDFs from meshes.
+
+pub mod grid;
